+++ service/engine/storage/cached-storage.cc
@@ -4,11 +4,10 @@
 
 #include "utils/log.h"
 #include "cached-storage.h"
-#include "cached-storage-page.h"
 
 namespace apollo {
 
-CachedStorage *CachedStorage::Init(std::string file_name, uint64_t page_size) {
+CachedStorage *CachedStorage::Init(std::string file_name, int page_size) {
   FILE *f = fopen(file_name.c_str(), "ab+");
 
   if (f == NULL) {
@@ -18,11 +17,11 @@ CachedStorage *CachedStorage::Init(std::string file_name, uint64_t page_size) {
   CachedStorage *storage = new CachedStorage(f, page_size);
 
   fseek(f, 0, SEEK_END);
-  uint64_t pages_count = ftell(f) / page_size + (ftell(f) % page_size ? 1 : 0);
+  long pages_count = ftell(f) / page_size + (ftell(f) % page_size ? 1 : 0);
   fseek(f, 0, SEEK_SET);
 
   for (int i = 0; i < pages_count; i++) {
-    storage->pages.push_back(CachedStoragePage::Load(f, page_size, i * page_size));
+    storage->pages.push_back(CachedStoragePage::Load(f, page_size, (uint64_t )i * page_size));
   }
 
   return storage;
@@ -45,25 +44,25 @@ StoragePage *CachedStorage::AllocatePage() {
   CachedStoragePage *page = CachedStoragePage::Load(
       this->file,
       this->page_size,
-      this->page_size * this->GetPagesCount());
+      (uint64_t)this->page_size * this->GetPagesCount());
   this->pages.push_back(page);
 
   return page;
 }
 
-StoragePage *CachedStorage::GetPage(uint64_t index) {
+StoragePage *CachedStorage::GetPage(int index) {
   return this->pages[index];
 }
 
-uint64_t CachedStorage::GetPagesCount() {
-  return this->pages.size();
+int CachedStorage::GetPagesCount() {
+  return (int)this->pages.size();
 }
 
 void CachedStorage::Flush() {
   fflush(this->file);
 }
 
-CachedStorage::CachedStorage(FILE *f, uint64_t page_size) {
+CachedStorage::CachedStorage(FILE *f, int page_size) {
   this->page_size = page_size;
   this->file = f;
 }
