+++ service/engine/storage/cached-storage.h
@@ -15,16 +15,16 @@ namespace apollo {
 class CachedStorage : public Storage {
  public:
   virtual ~CachedStorage();
-  static CachedStorage *Init(std::string file_name, uint64_t page_size);
+  static CachedStorage *Init(std::string file_name, int page_size);
 
   virtual StoragePage *AllocatePage();
-  virtual StoragePage *GetPage(uint64_t index);
-  virtual uint64_t GetPagesCount();
+  virtual StoragePage *GetPage(int index);
+  virtual int GetPagesCount();
   virtual void Flush();
  private:
-  CachedStorage(FILE *f, uint64_t page_size);
+  CachedStorage(FILE *f, int page_size);
 
-  uint64_t page_size;
+  int page_size;
   FILE *file;
   std::vector<CachedStoragePage *> pages;
 };
