+++ service/engine/storage/storage-page.h
@@ -14,9 +14,11 @@ typedef uint8_t volatile_t;
 class StoragePage {
  public:
   virtual ~StoragePage() {};
-  virtual void Write(uint64_t offset, void *source, uint64_t size) = 0;
-  virtual volatile_t *Read(uint64_t offset, uint64_t size) = 0;
-  virtual uint64_t GetPageSize() = 0;
+
+  virtual void Write(int offset, void *source, int size) = 0;
+  virtual volatile_t *Read(int offset, int size) = 0;
+
+  virtual int GetPageSize() = 0;
 };
 
 }
